finished new comparison method
