+++ docs/changelogs/changelog-v0.4.0.md
@@ -2,3 +2,4 @@
 
 ## What's New:
 * Ternary Operation for assigning values based on booleans
+* Moved comparison operations to new function
