+++ src/core/stsclasses.h
@@ -23,6 +23,7 @@ class sts;
 class stsvars;
 class stsfunc;
 
+// for the most part, it is a better idea to use functions outside of the sts class
 string striplit(string line);
 bool condition(sts *program, int *ln, std::vector<stsvars> vars);
 bool isint(string s);
