+++ tests/outputs/mathtest.sts.txt
@@ -1,0 +1,1 @@
+3 2 60 20
